Too many changes for one commit.
This commit tries to reduce the scope and complexity of the project.
I've tried to make one-off utils local, remove packages with one file
only, remove features that would be not used by me, rename things when
they are unclear to me.

Larger changes include how parsing of sources are managed by using
SourceFileRepository and the SourceFile sum-type.

THIS IS PROBABLY BROKEN